COMPRESS-147 yet another theoretical int overflow
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1548929 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/compressors/snappy/SnappyCompressorInputStream.java
@@ -231,7 +231,7 @@ public class SnappyCompressorInputStream extends CompressorInputStream {
                 offset = readOneByte();
                 offset |= readOneByte() << 8;
                 offset |= readOneByte() << 16;
-                offset |= readOneByte() << 24;
+                offset |= ((long) readOneByte()) << 24;
 
                 if (expandCopy(offset, length)) {
                     return;
